Move the `Result` class into a separate file
Closes #1560
